+++ .prettierrc
@@ -4,7 +4,7 @@
     "useTabs": false,
     "semi": true,
     "singleQuote": true,
-    "trailingComma": "es5",
+    "trailingComma": "all",
     "bracketSpacing": true,
     "jsxSingleQuote": false,
     "arrowParens": "always",
