+++ src/constants.ts
@@ -1,0 +1,6 @@
+export enum STATE {
+  S0 = 0,
+  S1 = 1,
+  S2 = 2,
+}
+export const ALPHABET: string[] = ['0', '1'];
