+++ tests/modThree.test.ts
@@ -1,23 +1,5 @@
-import {
-  validateInput,
-  transition,
-  modThree,
-  STATE,
-} from '../src/modThree';
-
-describe('validateInput', () => {
-  it('should not throw an error for valid input', () => {
-    expect(() => validateInput('01', ['0', '1'])).not.toThrow();
-    expect(() => validateInput('00', ['0', '1'])).not.toThrow();
-    expect(() => validateInput('11', ['0', '1'])).not.toThrow();
-  });
-
-  it('should throw an error for invalid input', () => {
-    expect(() => validateInput('02', ['0', '1'])).toThrow();
-    expect(() => validateInput('ab', ['0', '1'])).toThrow();
-    expect(() => validateInput('', ['0', '1'])).toThrow();
-  });
-});
+import { transition, modThree } from '../src/modThree';
+import { STATE } from '../src/constants';
 
 describe('transition', () => {
   it('should transition correctly from STATE S0', () => {
